Move SingletonAdmin to utils. Resolve #1480.
